+++ CHANGELOG.rst
@@ -2,6 +2,19 @@ Changelog
 =========
 
 
+Version v30.7.0
+----------------
+
+- Add composer and golang versions and also support gitlab native ranges for them
+
+
+Version v30.6.0
+----------------
+
+- Add support for gitlab in pypi, npm, gem version ranges
+- Subclass semver for nginx
+
+
 Version v30.5.1
 ----------------